% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_analysis.R
\name{new_asset_returns}
\alias{new_asset_returns}
\title{Create an asset_returns Object}
\usage{
new_asset_returns(
  data,
  date_col = "Date",
  return_col = "Return",
  asset_name = "Asset"
)
}
\arguments{
\item{data}{A data frame containing date and return columns.}

\item{date_col}{String. Name of the date column.}

\item{return_col}{String. Name of the return column.}

\item{asset_name}{String. Name of the asset (e.g., "SP500").}
}
\value{
An object of class "asset_returns".
}
\description{
Converts a data frame into an S3 object of class 'asset_returns'.
Performs defensive checks to ensure data validity.
}
\examples{
data(sp500)
my_asset <- new_asset_returns(sp500, date_col = "Date", return_col = "Return")
}
