% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_analysis.R
\name{plot.asset_returns}
\alias{plot.asset_returns}
\title{Plot Method for Asset Returns}
\usage{
\method{plot}{asset_returns}(x, type = "time", frequency = "daily", ...)
}
\arguments{
\item{x}{An object of class asset_returns.}

\item{type}{String. "time", "hist", or "qq".}

\item{frequency}{String. "daily", "monthly", or "yearly". Default is "daily".}

\item{...}{Additional arguments.}
}
\description{
Generates visualizations using ggplot2 with adjustable frequency.
}
