% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_analysis.R
\name{summary.asset_returns}
\alias{summary.asset_returns}
\title{Summary Method for Asset Returns}
\usage{
\method{summary}{asset_returns}(object, ...)
}
\arguments{
\item{object}{An object of class asset_returns.}

\item{...}{Additional arguments.}
}
\description{
Prints statistical moments (Mean, Volatility, Skewness, Kurtosis).
}
